Fixed bug in make_arff.py
There was a bug in make_arff.py: raw backslashes were being printed as, of course, raw backslashes, which is fine, but they are treated as escaping characters in .arff files. So, if '\"' just happed to occur naturally, it would be treated as a double-quote. If the double quote that's being escaped accidentally is the double quote that ends the instance text, then it's unclear where the .arff file ends. This will turn '\"' into '\\"', i.e., the backslash will be treated as a real backslash.

Also, the names of the attributes in the resulting .arff file have been changed to string_attribute and numeric_attribute since Weka seems to be complaining if the names are anything else.