Updated read_data_files.py with \b characters
Don't forget to put the \b characters at the beginning and end of the regular expressions: we only want to replace contractions that occur as words. \b is a zero-length word-boundary marker. So, '\bthis\b' would match "this" in "this is bad" but wouldn't match "this" in "his last name was Mathis".